Convert adpcr object to array

Converts adpcr object into the list of array-like matrices.

adpcr2panel(input, breaks = TRUE)

Arguments

input

object of the adpcr class.

breaks

if TRUE, the data is divided into intervals.

Value

A named list of length equal to the number of arrays in the input. Each element is a single array in matrix-like form, where dimensions are set exactly as in case of the real plate. Names of the list corresponds to the names of assays ("tnp" data) or runs (any other type of adpcr data).

The matrices contain values from array, either integers (when use_break is FALSE) or characters (when use_break is TRUE).

Examples

#generate data
ttest <- sim_adpcr(m = 400, n = 765, times = 20, pos_sums = FALSE,
                   n_panels = 3)

The assumed volume of partitions in each run is equal to 1.

The assumed volume uncertainty in each run is equal to 0.

#convert object into three arrays
arrays <- adpcr2panel(ttest)
length(arrays)

#> [1] 3
#> 

#print an array
arrays[[1]]

#>    1   2   3   4   5   6   7   8   9   10  11  12  13  14  15  16  17  18  19 
#> 1  "1" "0" "1" "0" "0" "1" "0" "0" "1" "0" "0" "2" "0" "1" "0" "1" "0" "1" "1"
#> 2  "0" "0" "2" "0" "0" "0" "0" "0" "0" "0" "0" "0" "0" "0" "0" "1" "0" "1" "0"
#> 3  "0" "0" "0" "0" "0" "0" "1" "0" "0" "1" "0" "1" "0" "1" "1" "0" "1" "0" "2"
#> 4  "2" "0" "0" "2" "1" "0" "1" "0" "0" "4" "0" "1" "0" "0" "1" "0" "1" "1" "0"
#> 5  "0" "3" "1" "0" "0" "0" "0" "1" "0" "0" "0" "0" "0" "0" "0" "0" "1" "1" "1"
#> 6  "1" "0" "0" "0" "2" "0" "0" "0" "0" "0" "0" "1" "1" "1" "1" "0" "0" "2" "1"
#> 7  "0" "2" "1" "0" "0" "1" "0" "2" "0" "0" "0" "1" "0" "0" "0" "0" "1" "0" "1"
#> 8  "0" "1" "0" "1" "0" "0" "0" "1" "0" "1" "0" "0" "0" "1" "0" "0" "0" "2" "0"
#> 9  "1" "0" "0" "0" "0" "0" "0" "1" "0" "0" "0" "0" "0" "0" "1" "0" "0" "2" "0"
#> 10 "0" "1" "0" "0" "2" "0" "1" "0" "0" "0" "0" "0" "1" "0" "0" "1" "0" "2" "0"
#> 11 "0" "1" "1" "1" "0" "0" "0" "0" "1" "2" "0" "0" "0" "0" "0" "0" "0" "0" "1"
#> 12 "1" "2" "0" "0" "0" "3" "1" "1" "0" "2" "1" "0" "1" "0" "0" "0" "0" "1" "0"
#> 13 "1" "1" "0" "0" "0" "1" "1" "1" "1" "2" "1" "2" "0" "0" "1" "1" "2" "0" "0"
#> 14 "1" "1" "0" "0" "1" "1" "0" "1" "0" "2" "1" "0" "0" "0" "1" "1" "0" "0" "2"
#> 15 "1" "0" "0" "1" "0" "0" "0" "2" "1" "4" "1" "1" "0" "0" "0" "0" "0" "1" "1"
#> 16 "1" "0" "1" "1" "0" "1" "0" "0" "0" "1" "0" "0" "0" "0" "0" "1" "1" "1" "0"
#> 17 "0" "0" "1" "0" "0" "0" "0" "0" "0" "0" "0" "0" "2" "0" "1" "0" "1" "0" "0"
#>    20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38 
#> 1  "0" "0" "0" "0" "1" "0" "0" "0" "1" "0" "1" "1" "0" "0" "0" "1" "2" "2" "0"
#> 2  "0" "1" "0" "0" "0" "0" "0" "0" "1" "0" "1" "1" "1" "1" "0" "1" "1" "0" "0"
#> 3  "0" "0" "1" "0" "0" "0" "0" "1" "1" "0" "0" "0" "0" "0" "0" "1" "1" "0" "2"
#> 4  "0" "0" "1" "1" "2" "0" "1" "0" "1" "0" "0" "1" "0" "1" "1" "0" "0" "0" "0"
#> 5  "0" "0" "0" "0" "0" "1" "0" "0" "1" "0" "1" "0" "0" "0" "0" "0" "1" "2" "2"
#> 6  "0" "0" "0" "0" "2" "1" "0" "1" "0" "0" "0" "1" "1" "1" "0" "1" "0" "0" "0"
#> 7  "0" "0" "0" "1" "0" "1" "0" "0" "0" "0" "0" "1" "1" "1" "1" "1" "1" "0" "0"
#> 8  "0" "0" "0" "1" "0" "1" "0" "0" "0" "0" "0" "0" "0" "0" "0" "1" "0" "0" "1"
#> 9  "0" "0" "0" "1" "0" "1" "2" "0" "0" "1" "1" "1" "1" "1" "0" "0" "1" "0" "0"
#> 10 "0" "1" "0" "0" "0" "0" "1" "1" "1" "0" "1" "0" "0" "0" "0" "0" "0" "0" "1"
#> 11 "1" "0" "1" "0" "0" "2" "0" "0" "0" "0" "0" "1" "0" "0" "0" "2" "0" "1" "3"
#> 12 "0" "1" "0" "0" "1" "1" "1" "1" "0" "1" "0" "0" "0" "0" "0" "0" "0" "1" "0"
#> 13 "2" "1" "0" "0" "2" "0" "0" "2" "1" "0" "2" "0" "0" "3" "1" "0" "0" "0" "0"
#> 14 "0" "0" "1" "0" "0" "0" "0" "1" "0" "0" "0" "0" "0" "0" "2" "2" "0" "0" "0"
#> 15 "0" "1" "0" "1" "0" "3" "0" "1" "0" "1" "1" "1" "1" "1" "0" "1" "0" "2" "0"
#> 16 "0" "0" "0" "1" "1" "0" "1" "1" "0" "1" "3" "2" "1" "0" "0" "0" "0" "0" "0"
#> 17 "0" "1" "0" "0" "1" "0" "1" "0" "0" "0" "1" "0" "2" "1" "0" "0" "0" "0" "2"
#>    39  40  41  42  43  44  45 
#> 1  "0" "1" "0" "0" "0" "0" "0"
#> 2  "1" "0" "0" "1" "0" "0" "2"
#> 3  "0" "1" "1" "0" "1" "0" "0"
#> 4  "0" "0" "0" "0" "2" "1" "0"
#> 5  "1" "2" "0" "0" "0" "1" "2"
#> 6  "0" "2" "0" "3" "0" "1" "0"
#> 7  "1" "1" "0" "1" "0" "0" "1"
#> 8  "2" "0" "1" "0" "1" "1" "0"
#> 9  "2" "0" "0" "0" "1" "0" "0"
#> 10 "3" "0" "0" "0" "2" "0" "1"
#> 11 "2" "1" "1" "0" "1" "2" "0"
#> 12 "0" "0" "0" "2" "1" "2" "0"
#> 13 "0" "0" "0" "0" "0" "0" "0"
#> 14 "0" "0" "3" "0" "0" "1" "2"
#> 15 "0" "2" "0" "0" "1" "1" "3"
#> 16 "0" "0" "1" "0" "0" "0" "0"
#> 17 "1" "2" "1" "0" "0" "1" "1"
#>