Convert adpcr to ppp

Converts adpcr object to the list of ppp.objects allowing spatial analysis.

adpcr2ppp(input, marks = TRUE, plot = FALSE)

Arguments

input

Object of the adpcr class containing data from one or more panels.

marks

If TRUE, marks values for non-empty partitions.

plot

If TRUE, array is plotted.

Value

A list containing objects with class ppp.object with the length equal to the number of arrays (minimum 1).

Details

Each array is independently converted by ppp function. marks attached to each point represent values contained by the adpcr object.

See also

ppp.object, ppp.

Examples

many_panels <- sim_adpcr(m = 400, n = 765, times = 1000, pos_sums = FALSE,
                   n_panels = 5)

The assumed volume of partitions in each run is equal to 1.

The assumed volume uncertainty in each run is equal to 0.

# Convert all arrays to ppp objects
adpcr2ppp(many_panels)

#> $Experiment1.1
#> Marked planar point pattern: 309 points
#> marks are numeric, of storage type  ‘double’
#> window: rectangle = [1, 45] x [1, 17] units
#> 
#> $Experiment1.2
#> Marked planar point pattern: 288 points
#> marks are numeric, of storage type  ‘double’
#> window: rectangle = [1, 45] x [1, 17] units
#> 
#> $Experiment1.3
#> Marked planar point pattern: 307 points
#> marks are numeric, of storage type  ‘double’
#> window: rectangle = [1, 45] x [1, 17] units
#> 
#> $Experiment1.4
#> Marked planar point pattern: 297 points
#> marks are numeric, of storage type  ‘double’
#> window: rectangle = [1, 45] x [1, 17] units
#> 
#> $Experiment1.5
#> Marked planar point pattern: 305 points
#> marks are numeric, of storage type  ‘double’
#> window: rectangle = [1, 45] x [1, 17] units
#> 
#> 

# Convert all arrays to ppp objects and get third plate
third_plate <- adpcr2ppp(many_panels)[[3]]

# Convert only third plate to ppp object
third_plate2 <- adpcr2ppp(extract_run(many_panels, 3))

# Check the class of a new object
class(third_plate2)

#> [1] "list"
#> 

# It's a list with the length 1. The third plate is a first element on this 
#list
class(third_plate2[[1]])

#> [1] "ppp"
#>