Bind dpcr objects

A convinient wrapper around cbind and rbind tailored specially for binding multiple objects containing results from digital PCR experiments.

bind_dpcr(input, ...)

Arguments

input

an object of class adpcr or dpcr or a list.

...

objects of class adpcr or dpcr. See Details. If input is a list, ignored.

Value

An object of class adpcr or dpcr, depending on the input.

Details

In case of adpcr or dpcr objects, bind_dpcr works analogously to cbind, but without recycling. In case on unequal length, shorter objects will be filled in with additional NA values. The original length is always preserved in n slot.

bind_dpcr automatically names binded experiments using format x.y, where x is number of object passed to function and y is a number of experiment in a given object.

Note

Because bind_dpcr calls do.call function, binding together at the same time more than 500 objects can lead to 'stack overflow' error.

See also

Opposite function: extract_run

Examples

bigger_array <- sim_adpcr(400, 765, 1000, pos_sums = FALSE, n_panels = 5)

The assumed volume of partitions in each run is equal to 1.

The assumed volume uncertainty in each run is equal to 0.

smaller_array <- sim_adpcr(100, 700, 1000, pos_sums = FALSE, n_panels = 3)

The assumed volume of partitions in each run is equal to 1.

The assumed volume uncertainty in each run is equal to 0.

bound_arrays <- bind_dpcr(bigger_array, smaller_array)

Different number of partitions. Shorter objects completed with NA values.

smaller_droplet <- sim_dpcr(m = 7, n = 20, times = 5, n_exp = 2)

The assumed volume of partitions in each run is equal to 1.

The assumed volume uncertainty in each run is equal to 0.

bigger_droplet <- sim_dpcr(m = 15, n = 25, times = 5, n_exp = 4)

The assumed volume of partitions in each run is equal to 1.

The assumed volume uncertainty in each run is equal to 0.

biggest_droplet <- sim_dpcr(m = 15, n = 35, times = 5, n_exp = 1)
bound_droplets <- bind_dpcr(smaller_droplet, bigger_droplet, biggest_droplet)

Different number of partitions. Shorter objects completed with NA values.