Class

Objects of this class are created by test_counts.

# S4 method for count_test
summary(object)

# S4 method for count_test
coef(object)

# S4 method for count_test
show(object)

# S4 method for count_test,ANY
plot(x, aggregate = FALSE, nice = TRUE)

Arguments

object

of class count_test.

x

object of class count_test.

aggregate

logical, if TRUE experiments are aggregated according to their group.

nice

logical, if TRUE a more aesthetically pleasing (but harder to customize) version of the plot is created.

Details

In case of the aggregated plot, mean confidence intervals for groups are presented as dashed lines.

Methods (by generic)

summary: Summary statistics of assigned groups.

coef: Extract coefficients of groups.

show: Print both group_coef and test_res.

plot: plots mean number of molecules per partition and its confidence intervals.

Slots

group_coef

"data.frame" containing experiments, groups to which they belong and calculated values of rate (lambda).

test_res

"matrix" containing result of multiple comparisions t-test.

model

"character" name of GLM used to compare experiments.

See also

test_counts.