Dispersion Test for Spatial Point Pattern in Array dPCR Based on Quadrat

Performs a test of Complete Spatial Randomness for each plate. This function is a wrapper around quadrat.test function working directly on the objects of adpcr.

test_panel(X, nx = 5, ny = 5, alternative = c("two.sided", "regular",
  "clustered"), method = c("Chisq", "MonteCarlo"), conditional = TRUE,
  nsim = 1999)

Arguments

X

Object of the adpcr class containing data from one or more panels.

nx

Number of quadrats in the x direction.

ny

Number of quadrats in the y direction.

alternative

character string (partially matched) specifying the alternative hypothesis.

method

character string (partially matched) specifying the test to use: either "Chisq" for the chi-squared test (the default), or "MonteCarlo" for a Monte Carlo test.

conditional

logical. Should the Monte Carlo test be conducted conditionally upon the observed number of points of the pattern? Ignored if method="Chisq".

nsim

The number of simulated samples to generate when method="MonteCarlo".

Value

A list of objects of class "htest" with the length equal to the number of plates (minimum 1).

Details

Under optimal conditions, the point pattern of dPCR events (e.g., positive droplet & negative droplets) should be randomly distrubuted over a planar chip. This function verifies this assumption using chi-square or Monte Carlo test. Arrays with non-random patterns should be checked for integrity.

Note

A similar result can be achived by using adpcr2ppp and quadrat.test. See Examples.

References

http://www.spatstat.org/

See also

quadrat.test.

Examples

many_panels <- sim_adpcr(m = 400, n = 765, times = 1000, pos_sums = FALSE,
                   n_panels = 5)

The assumed volume of partitions in each run is equal to 1.

The assumed volume uncertainty in each run is equal to 0.

test_panel(many_panels)

#> $Experiment1.1
#> 
#> 	Chi-squared test of CSR using quadrat counts
#> 	Pearson X2 statistic
#> 
#> data:  single_panel
#> X2 = 18.616, df = 24, p-value = 0.4558
#> alternative hypothesis: two.sided
#> 
#> Quadrats: 5 by 5 grid of tiles
#> 
#> $Experiment1.2
#> 
#> 	Chi-squared test of CSR using quadrat counts
#> 	Pearson X2 statistic
#> 
#> data:  single_panel
#> X2 = 18.336, df = 24, p-value = 0.4272
#> alternative hypothesis: two.sided
#> 
#> Quadrats: 5 by 5 grid of tiles
#> 
#> $Experiment1.3
#> 
#> 	Chi-squared test of CSR using quadrat counts
#> 	Pearson X2 statistic
#> 
#> data:  single_panel
#> X2 = 13.993, df = 24, p-value = 0.1064
#> alternative hypothesis: two.sided
#> 
#> Quadrats: 5 by 5 grid of tiles
#> 
#> $Experiment1.4
#> 
#> 	Chi-squared test of CSR using quadrat counts
#> 	Pearson X2 statistic
#> 
#> data:  single_panel
#> X2 = 25.63, df = 24, p-value = 0.7444
#> alternative hypothesis: two.sided
#> 
#> Quadrats: 5 by 5 grid of tiles
#> 
#> $Experiment1.5
#> 
#> 	Chi-squared test of CSR using quadrat counts
#> 	Pearson X2 statistic
#> 
#> data:  single_panel
#> X2 = 9.697, df = 24, p-value = 0.008626
#> alternative hypothesis: two.sided
#> 
#> Quadrats: 5 by 5 grid of tiles
#> 
#> 

#test only one plate
test_panel(extract_run(many_panels, 3))

#> $Experiment1.3
#> 
#> 	Chi-squared test of CSR using quadrat counts
#> 	Pearson X2 statistic
#> 
#> data:  single_panel
#> X2 = 13.993, df = 24, p-value = 0.1064
#> alternative hypothesis: two.sided
#> 
#> Quadrats: 5 by 5 grid of tiles
#> 
#> 

#do test_panel manually
require(spatstat)

Loading required package: spatstat

Loading required package: nlme

Loading required package: rpart

spatstat 1.46-1       (nickname: ‘Spoiler Alert’) 
For an introduction to spatstat, type ‘beginner’ 

Note: spatstat version 1.46-1 is out of date by more than 3 months; we recommend upgrading to the latest version.

Attaching package: ‘spatstat’

The following object is masked from ‘package:MASS’:

    area

ppp_data <- adpcr2ppp(many_panels)
lapply(ppp_data, function(single_panel) quadrat.test(single_panel))

#> $Experiment1.1
#> 
#> 	Chi-squared test of CSR using quadrat counts
#> 	Pearson X2 statistic
#> 
#> data:  single_panel
#> X2 = 18.616, df = 24, p-value = 0.4558
#> alternative hypothesis: two.sided
#> 
#> Quadrats: 5 by 5 grid of tiles
#> 
#> $Experiment1.2
#> 
#> 	Chi-squared test of CSR using quadrat counts
#> 	Pearson X2 statistic
#> 
#> data:  single_panel
#> X2 = 18.336, df = 24, p-value = 0.4272
#> alternative hypothesis: two.sided
#> 
#> Quadrats: 5 by 5 grid of tiles
#> 
#> $Experiment1.3
#> 
#> 	Chi-squared test of CSR using quadrat counts
#> 	Pearson X2 statistic
#> 
#> data:  single_panel
#> X2 = 13.993, df = 24, p-value = 0.1064
#> alternative hypothesis: two.sided
#> 
#> Quadrats: 5 by 5 grid of tiles
#> 
#> $Experiment1.4
#> 
#> 	Chi-squared test of CSR using quadrat counts
#> 	Pearson X2 statistic
#> 
#> data:  single_panel
#> X2 = 25.63, df = 24, p-value = 0.7444
#> alternative hypothesis: two.sided
#> 
#> Quadrats: 5 by 5 grid of tiles
#> 
#> $Experiment1.5
#> 
#> 	Chi-squared test of CSR using quadrat counts
#> 	Pearson X2 statistic
#> 
#> data:  single_panel
#> X2 = 9.697, df = 24, p-value = 0.008626
#> alternative hypothesis: two.sided
#> 
#> Quadrats: 5 by 5 grid of tiles
#> 
#>