Compare pooled digital PCR

Estimates mean number of template molecules per partition and concentration of sample from pooled replicates of experiments.

test_pooled(input, conf.level = 0.05)

Arguments

input

object of class adpcr or dpcr.

conf.level

confidence level of the intervals and groups.

Value

data frame with the number of rows equal to the number of experiments (not runs). The unit of concentration is the number template molecules per nanoliter (nL).

Note

This function was implemented using the code in supplemental materials in Dorazio, 2015 (see References).

References

Dorazio RM, Hunter ME, Statistical Models for the Analysis and Design of Digital Polymerase Chain Reaction (dPCR) Experiments. Analytical Chemistry 2015. 87(21): p.10886-10893

Examples

test_pooled(six_panels)

#>   sample_name          c      c.low       c.up       GOF
#> 1 Experiment1 0.01234183 0.00787226 0.01934904 0.9311255
#> 2 Experiment1 0.05034513 0.04038125 0.06276756 0.9311255
#> 3 Experiment2 0.12285367 0.10694519 0.14112860 0.9311255
#> 4 Experiment2 0.01234183 0.00787226 0.01934904 0.9311255
#> 5 Experiment3 0.05034513 0.04038125 0.06276756 0.9311255
#> 6 Experiment3 0.12285367 0.10694519 0.14112860 0.9311255
#>