Rate ratio test

The test for comparing ratio of two Poisson means: \(r = \frac{\lambda_1}{\lambda_2}\).

test_ratio(dpcr1, dpcr2, alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95)

Arguments

dpcr1

a (non-empty) numeric vector of data values of length 2 or more or an object of class adpcr or dpcr. See Details.

dpcr2

a (non-empty) numeric vector of data values of length 2 or more or an object of class adpcr or dpcr. See Details.

alternative

alternative hypothesis, must be one of: two.sided, greater or less.

conf.level

confidence level for the returned confidence interval.

Value

An object of class `htest' containing the following components:

Details

Objects dpcr1 and dpcr2 can be:

numeric vectors of length 2. The first element is assumed to be number of positive partitions and the second one to be the total number of partitions.

numeric vectors of length greater than 2. The length of vector is assumed to represent total number of partitions. Every element of the vector with value bigger than 0 is assumed to be a positive partitions.

adpcr objects with type tnp (total number of positive wells in panel) or nm (number of molecules per partition). dpcr objects with type tnp (total number of positive droplets) or nm (number of molecules per droplet).

Both dpcr1 and dpcr2 must have the same class. See Examples.

The ratio_test is a wrapper around rateratio.test function with custom input and output tailored specifically for digital PCR experiments.

References

Fay M.P. Two-sided exact tests and matching confidence intervals for discrete data R Journal 2 (1), 2010.

See also

See also poisson.test.

Examples

# Input values are numeric vectors representing dPCR experiments
x1 <- rpois(765, 1.1)
x2 <- rpois(765, 1.1)
test_ratio(x1, x2)

#> 
#> 	Exact Rate Ratio Test, assuming Poisson counts
#> 
#> data:  dPCR 1: positive partitions: 505; total partitions: 765.
#>        dPCR 2: positive partitions: 503; total partitions: 765.
#> p-value = 0.9749
#> alternative hypothesis: two.sided
#> 95 percent confidence interval:
#>  0.885601 1.138183
#> sample estimates:
#>         Lambda1         Lambda2 Lambda1/Lambda2 
#>        1.079194        1.071531        1.003976 
#> 
#> 

# Input values represent only number of positive partitions and total 
# partitions
x3 <- sum(rpois(765, 1.1) > 0)
x4 <- sum(rpois(765, 1.1) > 0)
test_ratio(c(x3, 765), c(x4, 765))

#> 
#> 	Exact Rate Ratio Test, assuming Poisson counts
#> 
#> data:  dPCR 1: positive partitions: 526; total partitions: 765.
#>        dPCR 2: positive partitions: 520; total partitions: 765.
#> p-value = 0.8771
#> alternative hypothesis: two.sided
#> 95 percent confidence interval:
#>  0.8943633 1.1440906
#> sample estimates:
#>         Lambda1         Lambda2 Lambda1/Lambda2 
#>        1.163412        1.138618        1.011538 
#> 
#> 

# It is possible to mix different types of input as long as they have 
# the same class
test_ratio(c(x3, 765), x1)

#> 
#> 	Exact Rate Ratio Test, assuming Poisson counts
#> 
#> data:  dPCR 1: positive partitions: 526; total partitions: 765.
#>        dPCR 2: positive partitions: 505; total partitions: 765.
#> p-value = 0.5334
#> alternative hypothesis: two.sided
#> 95 percent confidence interval:
#>  0.9200984 1.1792044
#> sample estimates:
#>         Lambda1         Lambda2 Lambda1/Lambda2 
#>        1.163412        1.079194        1.041584 
#> 
#> 

# The same is true for adpcr and dpcr objects.
x5 <- sim_adpcr(400, 1600, 100, pos_sums = TRUE, n_panels = 1)
x6 <- sim_adpcr(400, 1600, 100, pos_sums = FALSE, n_panels = 1)
test_ratio(x5, x6)

#> 
#> 	Exact Rate Ratio Test, assuming Poisson counts
#> 
#> data:  dPCR 1: positive partitions: 345; total partitions: 1600.
#>        dPCR 2: positive partitions: 359; total partitions: 1600.
#> p-value = 0.6242
#> alternative hypothesis: two.sided
#> 95 percent confidence interval:
#>  0.826576 1.117163
#> sample estimates:
#>         Lambda1         Lambda2 Lambda1/Lambda2 
#>       0.2428681       0.2540861       0.9610028 
#> 
#> 

x7 <- sim_dpcr(400, 1600, 100, pos_sums = TRUE, n_exp = 1)
x8 <- sim_dpcr(400, 1600, 100, pos_sums = FALSE, n_exp = 1)
test_ratio(x7, x8)

#> 
#> 	Exact Rate Ratio Test, assuming Poisson counts
#> 
#> data:  dPCR 1: positive partitions: 349; total partitions: 1600.
#>        dPCR 2: positive partitions: 355; total partitions: 1600.
#> p-value = 0.8505
#> alternative hypothesis: two.sided
#> 95 percent confidence interval:
#>  0.8456299 1.1428582
#> sample estimates:
#>         Lambda1         Lambda2 Lambda1/Lambda2 
#>       0.2460604       0.2508681       0.9830986 
#> 
#>